Nesting paragraph inside message div and giving it a class
This is less likely to break peoples custom styles since the origional div tag is still a div and I've added a class to it that didn't exist before.

The div just has a p tag inside it now instead of replacing by the p tag making it far less likely to break anyones custom styles.

By adding a class to the p tag, the only way that this would break peoples styles is if they customised the text by adding multiple p tags inside the text variable... although that's entirly possible :/